Trying to re-do how the platform is drawn
Using different drawing first-class functions, and swapping out which
one Platform.prototype.draw points to. Problem is, I don't change
platform.leftColor or .rightColor, so the bounce detection is off. I
need to find a better way to detect bounces, or just ditch the whole
thing and go back to the way it was.